Fix DefaultTasksRepository source preference - Give local source preference over remote source while fetching Tasks - Refactor and separate repository 'outbound request' (interface) from implementation
