+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/domain/usecase/ActivateTasksUseCase.kt
@@ -1,7 +1,7 @@
-package com.example.android.architecture.blueprints.todoapp.domain
+package com.example.android.architecture.blueprints.todoapp.domain.usecase
 
 import com.example.android.architecture.blueprints.todoapp.data.Task
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository
+import com.example.android.architecture.blueprints.todoapp.domain.repository.TasksRepository
 import com.example.android.architecture.blueprints.todoapp.util.wrapEspressoIdlingResource
 
 class ActivateTaskUseCase(
