+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/domain/usecase/DeleteTaskUseCase.kt
@@ -1,6 +1,6 @@
-package com.example.android.architecture.blueprints.todoapp.domain
+package com.example.android.architecture.blueprints.todoapp.domain.usecase
 
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository
+import com.example.android.architecture.blueprints.todoapp.domain.repository.TasksRepository
 import com.example.android.architecture.blueprints.todoapp.util.wrapEspressoIdlingResource
 
 class DeleteTaskUseCase(
