+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/domain/usecase/GetTaskUseCase.kt
@@ -1,8 +1,8 @@
-package com.example.android.architecture.blueprints.todoapp.domain
+package com.example.android.architecture.blueprints.todoapp.domain.usecase
 
 import com.example.android.architecture.blueprints.todoapp.data.Result
 import com.example.android.architecture.blueprints.todoapp.data.Task
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository
+import com.example.android.architecture.blueprints.todoapp.domain.repository.TasksRepository
 import com.example.android.architecture.blueprints.todoapp.util.wrapEspressoIdlingResource
 
 class GetTaskUseCase(
