+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/domain/usecase/GetTasksUseCase.kt
@@ -1,9 +1,9 @@
-package com.example.android.architecture.blueprints.todoapp.domain
+package com.example.android.architecture.blueprints.todoapp.domain.usecase
 
 import com.example.android.architecture.blueprints.todoapp.data.Result
 import com.example.android.architecture.blueprints.todoapp.data.Result.Success
 import com.example.android.architecture.blueprints.todoapp.data.Task
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository
+import com.example.android.architecture.blueprints.todoapp.domain.repository.TasksRepository
 import com.example.android.architecture.blueprints.todoapp.tasks.TasksFilterType
 import com.example.android.architecture.blueprints.todoapp.tasks.TasksFilterType.ACTIVE_TASKS
 import com.example.android.architecture.blueprints.todoapp.tasks.TasksFilterType.ALL_TASKS
