+++ app/src/sharedTest/java/com/example/android/architecture/blueprints/todoapp/util/TasksRepositoryExt.kt
@@ -16,7 +16,7 @@
 package com.example.android.architecture.blueprints.todoapp.util
 
 import com.example.android.architecture.blueprints.todoapp.data.Task
-import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepository
+import com.example.android.architecture.blueprints.todoapp.domain.repository.TasksRepository
 import kotlinx.coroutines.runBlocking
 
 /**
